+++ plone/app/discussion/browser/javascripts/moderation.js
@@ -24,10 +24,9 @@ require([  // jshint ignore:line
     //#JSCOVERAGE_IF 0
 
     /**************************************************************************
-     * Window Load Function: Executes when complete page is fully loaded,
-     * including all frames,
+     * Document Ready Function: Executes when DOM is ready.
      **************************************************************************/
-    $(window).load(function () {
+    $(document).ready(function () {
 
         /**********************************************************************
          * Delete a single comment.